requirements.txt
pefile>=2021.9.3
psutil>=5.8.0
yara-python>=4.2.0
scapy>=2.4.5
pyshark>=0.4.3
watchdog>=2.1.6
colorama>=0.4.4
python-magic>=0.4.24
tabulate>=0.8.9
jinja2>=3.0.1
termcolor>=1.1.0
ipaddress>=1.0.23
# Linux-specific dependencies (optional)
# libpcap-dev  # For packet capture on Linux (install via package manager)
# tcpdump      # For network analysis (install via package manager)